+++ LinearAlgebra/LinearAlgebra_Step11 Scan Conversion3 FillTriangle/KVectorUtil.cpp
@@ -88,8 +88,8 @@ void KVectorUtil::PutPixel(HDC hdc, int x, int y, Gdiplus::Color color)
     {
         KVector2 v0 = g_basis2.Transform(v[i]);
         v0 = g_screenCoordinate.Transform(v0);
-        point[i].X = v0.x;
-        point[i].Y = v0.y;
+        point[i].X = (int)v0.x;
+        point[i].Y = (int)v0.y;
     }
 
     Gdiplus::SolidBrush  brush(color);
@@ -101,8 +101,8 @@ void KVectorUtil::_ScanLineLow(HDC hdc, int x0, int y0, int x1, int y1, Gdiplus:
 {
     auto sign = [](float delta){ return delta > 0.f ? 1.0f : -1.0f; };
 
-    float deltax = x1 - x0;
-    float deltay = y1 - y0;
+    float deltax = float(x1 - x0);
+    float deltay = float(y1 - y0);
     float deltaerr = abs(deltay / deltax); // Assume deltax != 0 (line is not vertical),
     // note that this division needs to be done in a way that preserves the fractional part
     float error = 0.0f; // No error at start
@@ -113,7 +113,7 @@ void KVectorUtil::_ScanLineLow(HDC hdc, int x0, int y0, int x1, int y1, Gdiplus:
         error = error + deltaerr;
         if (error >= 0.5f)
         {
-            y = y + sign(deltay) * 1.0f;
+            y = y + int(sign(deltay) * 1.0f);
             error = error - 1.0f;
         }
     }
@@ -123,8 +123,8 @@ void KVectorUtil::_ScanLineHigh(HDC hdc, int x0, int y0, int x1, int y1, Gdiplus
 {
     auto sign = [](float delta){ return delta > 0.f ? 1.0f : -1.0f; };
 
-    float deltax = x1 - x0;
-    float deltay = y1 - y0;
+    float deltax = float(x1 - x0);
+    float deltay = float(y1 - y0);
     float deltaerr = abs(deltax / deltay); // Assume deltax != 0 (line is not vertical),
     // note that this division needs to be done in a way that preserves the fractional part
     float error = 0.0f; // No error at start
@@ -135,7 +135,7 @@ void KVectorUtil::_ScanLineHigh(HDC hdc, int x0, int y0, int x1, int y1, Gdiplus
         error = error + deltaerr;
         if (error >= 0.5f)
         {
-            x = x + sign(deltax) * 1.0f;
+            x = x + int(sign(deltax) * 1.0f);
             error = error - 1.0f;
         }
     }
