+++ LinearAlgebra/LinearAlgebra_Step12 Colored Cube/KPolygon.cpp
@@ -68,15 +68,19 @@ void ScanDrawIndexedPrimitive( HDC hdc
         // draw triangle
         if(isVisible == true)
         {
-            KVectorUtil::FillTriangle( hdc, m_vertexBuffer[i0].x, m_vertexBuffer[i0].y, color_
-                , m_vertexBuffer[i1].x, m_vertexBuffer[i1].y, color_
-                , m_vertexBuffer[i2].x, m_vertexBuffer[i2].y, color_ );
-            KVectorUtil::ScanLineSegment( hdc, m_vertexBuffer[i0].x, m_vertexBuffer[i0].y, KRgb(0, 0, 0)
-                , m_vertexBuffer[i1].x, m_vertexBuffer[i1].y, KRgb( 0, 0, 0 ), nullptr );
-            KVectorUtil::ScanLineSegment( hdc, m_vertexBuffer[i1].x, m_vertexBuffer[i1].y, KRgb(0, 0, 0)
-                , m_vertexBuffer[i2].x, m_vertexBuffer[i2].y, KRgb( 0, 0, 0 ), nullptr );
-            KVectorUtil::ScanLineSegment( hdc, m_vertexBuffer[i2].x, m_vertexBuffer[i2].y, KRgb(0, 0, 0)
-                , m_vertexBuffer[i0].x, m_vertexBuffer[i0].y, KRgb( 0, 0, 0 ), nullptr );
+            KVectorUtil::FillTriangle( hdc
+                , (int)m_vertexBuffer[i0].x, (int)m_vertexBuffer[i0].y, color_
+                , (int)m_vertexBuffer[i1].x, (int)m_vertexBuffer[i1].y, color_
+                , (int)m_vertexBuffer[i2].x, (int)m_vertexBuffer[i2].y, color_ );
+            KVectorUtil::ScanLineSegment( hdc
+                , (int)m_vertexBuffer[i0].x, (int)m_vertexBuffer[i0].y, KRgb(0, 0, 0)
+                , (int)m_vertexBuffer[i1].x, (int)m_vertexBuffer[i1].y, KRgb( 0, 0, 0 ), nullptr );
+            KVectorUtil::ScanLineSegment( hdc
+                , (int)m_vertexBuffer[i1].x, (int)m_vertexBuffer[i1].y, KRgb(0, 0, 0)
+                , (int)m_vertexBuffer[i2].x, (int)m_vertexBuffer[i2].y, KRgb( 0, 0, 0 ), nullptr );
+            KVectorUtil::ScanLineSegment( hdc
+                , (int)m_vertexBuffer[i2].x, (int)m_vertexBuffer[i2].y, KRgb(0, 0, 0)
+                , (int)m_vertexBuffer[i0].x, (int)m_vertexBuffer[i0].y, KRgb( 0, 0, 0 ), nullptr );
         }
         // advance to next primitive
         counter += 3;
